+++ packages/@glimmer/application/src/references.ts
@@ -1,4 +1,6 @@
+import { DEBUG } from '@glimmer/env';
 import { dict } from '@glimmer/util';
+import { Option } from '@glimmer/interfaces';
 import {
   PathReference,
   ConstReference,
@@ -14,12 +16,102 @@ import {
   update,
   createTag,
   dirty,
+  track,
+  consume,
 } from '@glimmer/validator';
 import {
   ConditionalReference as GlimmerConditionalReference,
   PrimitiveReference,
 } from '@glimmer/runtime';
-import { consume, trackProperty } from '@glimmer/tracking';
+
+export function trackProperty<T = unknown>(obj: {}, key: string, throwError = defaultErrorThrower): [T, Tag] {
+  if (DEBUG && typeof obj === 'object') {
+    installDevModeErrorInterceptor(obj, key, throwError);
+  }
+  let value;
+  const tag = track(() => { value = obj[key]; });
+  return [value, tag];
+}
+
+export class UntrackedPropertyError extends Error {
+  static for(obj: any, key: string): UntrackedPropertyError {
+    return new UntrackedPropertyError(
+      obj,
+      key,
+      `The property '${key}' on ${obj} was changed after being rendered. If you want to change a property used in a template after the component has rendered, mark the property as a tracked property with the @tracked decorator.`
+    );
+  }
+
+  constructor(public target: any, public key: string, message: string) {
+    super(message);
+  }
+}
+
+/**
+ * Function that can be used in development mode to generate more meaningful
+ * error messages.
+ */
+export interface UntrackedPropertyErrorThrower {
+  (obj: any, key: string): void;
+}
+
+function defaultErrorThrower(obj: any, key: string): UntrackedPropertyError {
+  throw UntrackedPropertyError.for(obj, key);
+}
+
+/**
+ * In development mode only, we install an ad hoc setter on properties where a
+ * tag is requested (i.e., it was used in a template) without being tracked. In
+ * cases where the property is set, we raise an error.
+ */
+function installDevModeErrorInterceptor(
+  obj: object,
+  key: string,
+  throwError: UntrackedPropertyErrorThrower
+) {
+  let target = obj;
+  let descriptor: Option<PropertyDescriptor> = null;
+
+  // Find the descriptor for the current property. We may need to walk the
+  // prototype chain to do so. If the property is undefined, we may never get a
+  // descriptor here.
+  let hasOwnDescriptor = true;
+  while (target) {
+    descriptor = Object.getOwnPropertyDescriptor(target, key);
+    if (descriptor) {
+      break;
+    }
+    hasOwnDescriptor = false;
+    target = Object.getPrototypeOf(target);
+  }
+
+  // If possible, define a property descriptor that passes through the current
+  // value on reads but throws an exception on writes.
+  if (descriptor) {
+    // Only install the interceptor if it's a simple (non-getter) property and
+    // the existing property is able to be configured.
+    if (descriptor.hasOwnProperty('value') && (descriptor.configurable || !hasOwnDescriptor)) {
+      Object.defineProperty(obj, key, {
+        configurable: descriptor.configurable,
+        enumerable: descriptor.enumerable,
+
+        get() {
+          return descriptor.value;
+        },
+
+        set() {
+          throwError(this, key);
+        },
+      });
+    }
+  } else {
+    Object.defineProperty(obj, key, {
+      set() {
+        throwError(this, key);
+      },
+    });
+  }
+}
 
 export abstract class ComponentPathReference<T> implements PathReference<T> {
   abstract value(): T;
