+++ packages/@glimmer/tracking/index.ts
@@ -1,10 +1,1 @@
-export {
-  tracked,
-  setPropertyDidChange,
-  metaFor,
-  trackedGet,
-  trackProperty,
-  track,
-  consume,
-  UntrackedPropertyError,
-} from './src/tracked';
+export { tracked, setPropertyDidChange } from './src/tracked';
