+++ test/types/tracking-test.ts
@@ -1,0 +1,13 @@
+import * as tracking from '@glimmer/tracking';
+import { hasExactKeys } from './utils';
+
+hasExactKeys<{
+  setPropertyDidChange: unknown,
+  tracked: unknown,
+}>()(tracking);
+
+// $ExpectType (cb: () => void) => void
+tracking.setPropertyDidChange;
+
+// $ExpectType PropertyDecorator
+tracking.tracked;
